+++ src/predictor.py
@@ -1,3 +1,5 @@
+# FastAPIを使ったオンライン推論サーバーのエントリーポイント
+# 起動時にS3からモデルをロードし、/predict エンドポイントで広告クリック率を予測する
 import json
 import logging
 import os
@@ -19,6 +21,8 @@ from mlops.predictor import AdRequest
 logger = logging.getLogger(__name__)
 
 
+# FastAPIアプリケーションの起動・終了時の処理を管理するライフスパンコンテキストマネージャ
+# アプリ起動時にモデルと特徴量ストアを初期化し、各リクエストから参照できるよう yield で渡す
 @asynccontextmanager
 async def lifespan(app: FastAPI) -> AsyncGenerator:
     current_time = datetime.now()
@@ -26,18 +30,21 @@ async def lifespan(app: FastAPI) -> AsyncGenerator:
     artifact = Artifact(version=version, job_type="predictor")
     set_logger_config(log_file_path=artifact.file_path("log.txt"))
 
+    # 環境変数からモデル名・バージョン・特徴量バージョンを取得する（デフォルト値付き）
     # Setup for model config
     model_name = os.getenv("MODEL_NAME", "sgd_classifier_ctr")
     model_version = os.getenv("MODEL_VERSION", "latest")
     feature_version = os.getenv("FEATURE_VERSION", "latest")
     logger.info(f"Configure {model_name=}, {model_version=}, {feature_version=}")
 
+    # model_version が "latest" の場合は DynamoDB のモデルレジストリから最新バージョンを取得する
     if model_version == "latest":
         latest_model_version = get_latest_model_version(table=MODEL_REGISTRY_DYNAMODB_TABLE, model=model_name)
         if latest_model_version is None:
             raise ValueError("No latest model found")
         model_version = latest_model_version
 
+    # DynamoDB からモデルの S3 パス（s3_key）を取得する
     model_s3_key = get_model_s3_key(table=MODEL_REGISTRY_DYNAMODB_TABLE, model=model_name, version=model_version)
     if model_s3_key is None:
         raise ValueError("Model S3 key not found")
@@ -46,9 +53,11 @@ async def lifespan(app: FastAPI) -> AsyncGenerator:
     #  S3 からモデルファイルをダウンロード
     model = model_config.model_class.from_pretrained(s3_key=model_s3_key)
 
+    # DynamoDB のオンライン特徴量ストアへの接続クライアントを初期化する
     online_feature_store = OnlineFeatureStoreDynamoDB(table=FEATURE_DYNAMODB_TABLE, version=feature_version)
 
     # 論結果をログとして記録するために、DataFrame に列を追加
+    # 予測結果・モデル情報・タイムスタンプを JSON として標準出力に書き出す
     def log_prediction(df: pd.DataFrame, prediction: float) -> None:
         df = df.assign(
             prediction=prediction,
@@ -59,6 +68,7 @@ async def lifespan(app: FastAPI) -> AsyncGenerator:
         )
         print(json.dumps(df.iloc[0].to_dict()))
 
+    # yield によって起動時に初期化したオブジェクトを request.state 経由で各エンドポイントに渡す
     yield {
         "model": model,
         "model_config": model_config,
@@ -70,28 +80,37 @@ async def lifespan(app: FastAPI) -> AsyncGenerator:
 app = FastAPI(lifespan=lifespan)
 
 
+# 広告クリック率を予測するエンドポイント
+# リクエストのユーザーID に基づき DynamoDB から特徴量を取得し、モデルで予測確率を返す
 @app.post("/predict")
 async def predict(ad_request: AdRequest, request: Request) -> dict[str, str | float]:
+    # リクエストボディを DataFrame に変換する（1行のみ）
     df = pd.DataFrame([ad_request.model_dump()])
 
+    # DynamoDB からユーザーの過去行動特徴量を取得して DataFrame に結合する
     # Get user feature from DynamoDB
     user_feature = request.state.online_feature_store.get_impression_feature(user_id=ad_request.user_id)
     df = df.assign(**user_feature)
 
+    # インプレッション時刻から時間・日・曜日の特徴量を追加する
     # Add impression time features
     df = add_impression_time_feature(df, "logged_at")
 
+    # モデルスキーマに従い、欠損値補完とデータ型変換を適用する
     # Fill missing values and convert data types
     for schema in request.state.model_config.schemas:
         if schema.name not in df.columns:
-            df[schema.name] = np.nan
-        df[schema.name] = df[schema.name].fillna(schema.null_value)
-        df[schema.name] = df[schema.name].astype(schema.dtype)
+            df[schema.name] = np.nan  # スキーマに定義されているが DataFrame に存在しない列は NaN で初期化
+        df[schema.name] = df[schema.name].fillna(schema.null_value)  # null_value で欠損補完
+        df[schema.name] = df[schema.name].astype(schema.dtype)  # 指定データ型にキャスト
+    # スキーマ定義の順序に従ってカラムを並び替える
     df = df[[schema.name for schema in request.state.model_config.schemas]]
 
+    # モデルによる予測確率を計算する（クリック率の予測値）
     # Get prediction
     prediction = request.state.model.predict_proba(df)
 
+    # 予測結果をログに記録する
     request.state.log_prediction(df, float(prediction))
 
     return dict(
@@ -100,11 +119,13 @@ async def predict(ad_request: AdRequest, request: Request) -> dict[str, str | fl
     )
 
 
+# ヘルスチェックエンドポイント（ロードバランサーやコンテナオーケストレーターからの死活監視用）
 @app.get("/healthcheck")
 async def healthcheck() -> dict[str, str]:
     return {"health": "ok"}
 
 
+# uvicorn でサーバーを起動するエントリーポイント関数
 def main() -> None:
     uvicorn.run("predictor:app", host="0.0.0.0", port=8080, reload=True)
